requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
CacheControl==0.12.11
cachetools==5.3.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cryptography==41.0.6
cssselect==1.2.0
cycler==0.11.0
Deprecated==1.2.13
exceptiongroup==1.1.1
fake-useragent==1.1.3
feedparser==6.0.10
firebase-admin==6.1.0
Flask==2.2.5
Flask-JWT-Extended==4.4.4
Flask-WTF==1.1.1
finnhub-python==2.4.18
fonttools==4.43.0
frozendict==2.3.7
frozenlist==1.3.3
gcloud==0.18.3
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpcio==1.54.0
grpcio-status==1.54.0
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
httpcore==0.17.0
httplib2==0.22.0
httpx==0.24.0
idna==3.4
importlib-metadata==6.5.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.2.0
jwcrypto==1.5.1
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
mplfinance==0.12.9b7
msgpack==1.0.5
multidict==6.0.4
multitasking==0.0.11
nltk==3.8.1
numpy==1.24.2
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.27.8
outcome==1.2.0
packaging==23.1
pandas==1.3.5
parse==1.19.0
Pillow==10.0.1
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.1
pyee==8.2.2
PyJWT==2.6.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==2.0.0
Pyrebase4==4.6.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jwt==4.0.0
pytz==2023.3
regex==2023.3.23
requests==2.31.0
requests-html==0.10.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rsa==4.9
selenium==4.9.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
textblob==0.17.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.13.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
w3lib==2.1.1
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.3.8
wrapt==1.15.0
wsproto==1.2.0
WTForms==3.0.1
yahoo-fin==0.8.9.1
yahoofinancials==1.14
yarl==1.9.1
yfinance==0.2.36
zipp==3.15.0
